Extract demo script into separate file
