+++ nuimo_dbus.py
@@ -1,5 +1,3 @@
-#!/usr/bin/env python3
-
 import dbus
 import dbus.mainloop.glib
 import functools
@@ -196,6 +194,7 @@ class NuimoLedMatrix:
     def __init__(self, string):
         string = "{:<81}".format(string[:81])
         self.leds = [c not in [' ', '0'] for c in string]
+
 
 class NuimoController(GattDevice):
     NUIMO_SERVICE_UUID                    = "f29b1525-cb19-40f3-be5c-7241ecb82fd2"
@@ -439,54 +438,3 @@ class NuimoControllerPrintListener(NuimoControllerListener):
     def print(self, string):
         print("Nuimo controller " + self.controller.mac_address + " " + string)
 
-
-class NuimoControllerTestListener(NuimoControllerPrintListener):
-    def __init__(self, controller):
-        super().__init__(controller)
-
-    def disconnected(self):
-        super().disconnected()
-
-        # Reconnect as soon as Nuimo was disconnected
-        # TODO: Only reconnect if `disconnect` was not called – add an error parameter to this callback
-        print("Disconnected, reconnecting...")
-        self.controller.connect()
-
-    def received_gesture_event(self, event):
-        super().received_gesture_event(event)
-        self.controller.display_matrix(NuimoLedMatrix(
-            "*        "
-            " *       "
-            "  *      "
-            "   *     "
-            "    *    "
-            "     *   "
-            "      *  "
-            "       * "
-            "        *"))
-
-
-class NuimoControllerManagerPrintListener(NuimoControllerManagerListener):
-    def controller_discovered(self, controller):
-        print("Discovered Nuimo controller", controller.mac_address)
-
-
-if __name__ == '__main__':
-    dbus.mainloop.glib.DBusGMainLoop(set_as_default=True)
-
-    #controller = NuimoController(adapter_name="hci0", mac_address="FC:52:6E:8E:87:06")
-    #controller = NuimoController(adapter_name="hci0", mac_address="C4:54:31:CD:DD:07")
-
-    #controller.listener = NuimoControllerTestListener(controller=controller)
-
-    #print("Connected:", controller.is_connected())
-    #controller.disconnect()
-    #controller.connect()
-
-    controller_manager = NuimoControllerManager(adapter_name="hci0")
-    controller_manager.listener = NuimoControllerManagerPrintListener()
-    controller_manager.start_discovery()
-
-    print("Entering main loop. Exit with Ctrl+C")
-    mainloop = GObject.MainLoop()
-    mainloop.run()
